Creating module location.js and refactoring useEffect in to the sperate moudle
